Add reference implementation of SigMF archive format from issue #15
If a SigMFFile holds valid metadata and a data_file path, then calling
the `archive` method tars metadata and data according to the spec, and
returns the path to the archive.

This commit also adds a function: `sigmffile.fromarchive`. If passed a
valid sigmf archive and a path, extract the archive to that path and
return a SigMFFile object holding the archive's metadata and pointing
to the extracted `data_file`.