Require metadata to be directly before the data member.
Requiring the metadata to come before all other information allows you
to process the archive's data without actually unpacking it to the
filesystem or copy all data into meomry. This idea is shamelessly stolen
from how we mange members inside the Debian `.deb` file format

The irony in `meta` strictly meaning "after" or "beyond" is not lost on
the author.